Implement weekly report properly
- Weekly report now compares the last two weeks
- Extract more generic stuff to report base class
- Add some documentation
